Implemented initial configuration functionality
